requirements.txt
numpy>=1.23,<2.0
scipy>=1.10,<2.0
pandas>=1.5,<3.0
matplotlib>=3.7,<4.0
scikit-image>=0.22,<0.23
scikit-learn>=1.3,<2.0
numba>=0.58,<1.0
tqdm>=4.66,<5.0
pyyaml>=6.0,<7.0
opencv-python>=4.8,<5.0
natsort>=8.4,<9.0
loguru>=0.7,<1.0
h5py>=3.9,<4.0
packaging>=23.2,<24.0
PySide6>=6.5,<7.0

# --- EBSD / Kikuchi pattern analysis ----------------------------------------
hyperspy>=2.2,<3.0          # new API with hs.plot.markers.Texts :contentReference[oaicite:0]{index=0}
kikuchipy>=0.11,<0.12       # requires HyperSpy ≥ 2.2 :contentReference[oaicite:1]{index=1}
orix>=0.13,<0.14            # orientation-crystallography helpers (optional)
pyxem>=0.18,<0.19           # 4D-STEM; harmless if unused

# --- GUI / interactive extras ----------------------------------------------
#pyqt5>=5.15,<5.16           # needed for TraitsUI viewer
#ipywidgets>=8.1,<9.0
#traits>=6.4,<7.0
#traitlets>=5.9,<6.0

# --- notebooks / dev convenience -------------------------------------------
#jupyterlab>=4.0,<5.0
#notebook>=7.1,<8.0
